feat: add decimal precision

+++ participantes/emiliosheinz-go/init.sql
@@ -5,7 +5,7 @@ CREATE TYPE processor AS ENUM ('default', 'fallback');
 CREATE TABLE payments (
   id SERIAL PRIMARY KEY,
   correlation_id VARCHAR(255) NOT NULL,
-  amount DOUBLE PRECISION NOT NULL,
+  amount DECIMAL NOT NULL,
   processed_at TIMESTAMP NOT NULL,
   processed_by processor NOT NULL
 );
